requirements.txt
# 📦 KanterMator + Sylvie - Dependencies
# Phase 1.3 & 2.1 - Backend Core

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Google Workspace APIs
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Database
psycopg2-binary==2.9.8
sqlalchemy==2.0.23
alembic==1.12.1

# Scheduling & Background Tasks
celery==5.3.4
redis==5.0.1
schedule==1.2.0

# IA Hybride - OpenAI + Google Gemini pour Sylvie
openai==1.3.6
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Ajout des dépendances pour Sylvie
openai>=1.0.0
asyncio-mqtt>=0.16.1
python-multipart>=0.0.6
websockets>=11.0.3

